// src/src/types/types.ts
export enum EBookingDriverState {
  Considering = 1,
  Canceled,
  Performer,
  Arrived,
  Started,
  Finished
}
// ...
export enum EUserRoles {
  Client = 1,
  Driver,
  Administrator,
  Agent
}
// ...
export enum EStatuses {
  Default,
  Loading,
  Success,
  Fail,
  Whatsapp,
  Warning,
}
// ...
export interface ILanguage {
  id: number
  iso: string
  logo: string
  native: string

  [language: string]: string | number
}